# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Couple Budget Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: month
          in: query
          schema: { type: ["string", "null"] }
          description: "Month filter in YYYY-MM format"
        - name: startDate
          in: query
          schema: { type: ["string", "null"] }
        - name: endDate
          in: query
          schema: { type: ["string", "null"] }
        - name: partnerId
          in: query
          schema: { type: ["integer", "null"] }
        - name: categoryId
          in: query
          schema: { type: ["integer", "null"] }
        - name: accountId
          in: query
          schema: { type: ["integer", "null"] }
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [income, expense, null]
        - name: ownership
          in: query
          schema:
            type: ["string", "null"]
            enum: [shared, personal, null]
        - name: isRecurring
          in: query
          schema: { type: ["boolean", "null"] }
        - name: search
          in: query
          schema: { type: ["string", "null"] }
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
            enum: [date, amount, merchant, category, null]
        - name: sortDir
          in: query
          schema:
            type: ["string", "null"]
            enum: [asc, desc, null]
        - name: limit
          in: query
          schema: { type: ["integer", "null"] }
        - name: offset
          in: query
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: Paginated list of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single transaction
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Budgets ───────────────────────────────────────────────────────────────
...
  /savings-goals:
    get:
      operationId: getSavingsGoals
      tags: [savings]
      summary: List all savings goals
      responses:
        "200":
          description: List of savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsGoal"
    post:
      operationId: createSavingsGoal
      tags: [savings]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalInput"
      responses:
        "201":
          description: Created savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
  /savings-goals/{id}:
    patch:
      operationId: updateSavingsGoal
      tags: [savings]
      summary: Update a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalUpdate"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
    delete:
      operationId: deleteSavingsGoal
      tags: [savings]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /savings-goals/{id}/contribute:
    post:
      operationId: contributeSavingsGoal
      tags: [savings]
      summary: Add a contribution to a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsContribution"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"

  # ── Dashboard / Analytics (wow endpoints) ─────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, date, amount, merchant, type, ownership, paidById, categoryId]
      properties:
        id:            { type: integer }
        date:          { type: string }
        amount:        { type: number }
        merchant:      { type: string }
        notes:         { type: ["string", "null"] }
        type:          { type: string, enum: [income, expense] }
        ownership:     { type: string, enum: [shared, personal] }
        splitType:     { type: string, enum: [fifty_fifty, custom, personal, settle_later] }
        splitRatio:    { type: ["number", "null"] }
        paidById:      { type: integer }
        categoryId:    { type: integer }
        accountId:     { type: ["integer", "null"] }
        isRecurring:   { type: boolean }
        categoryName:  { type: ["string", "null"] }
        categoryColor: { type: ["string", "null"] }
        paidByName:    { type: ["string", "null"] }
    TransactionInput:
      type: object
      required: [date, amount, merchant, type, ownership, splitType, paidById, categoryId]
      properties:
        date:        { type: string }
        amount:      { type: number }
        merchant:    { type: string }
        notes:       { type: ["string", "null"] }
        type:        { type: string, enum: [income, expense] }
        ownership:   { type: string, enum: [shared, personal] }
        splitType:   { type: string, enum: [fifty_fifty, custom, personal, settle_later] }
        splitRatio:  { type: ["number", "null"] }
        paidById:    { type: integer }
        categoryId:  { type: integer }
        accountId:   { type: ["integer", "null"] }
        isRecurring: { type: boolean }
    TransactionUpdate:
      type: object
      properties:
        date:        { type: string }
        amount:      { type: number }
        merchant:    { type: string }
        notes:       { type: ["string", "null"] }
        type:        { type: string, enum: [income, expense] }
        ownership:   { type: string, enum: [shared, personal] }
        splitType:   { type: string, enum: [fifty_fifty, custom, personal, settle_later] }
        splitRatio:  { type: ["number", "null"] }
        paidById:    { type: integer }
        categoryId:  { type: integer }
        accountId:   { type: ["integer", "null"] }
        isRecurring: { type: boolean }
    TransactionList:
      type: object
      required: [data, total]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Transaction" } }
        total: { type: integer }

    # ── Budgets ───────────────────────────────────────────────────────────────
...
    SavingsGoal:
      type: object
      required: [id, name, targetAmount, currentAmount, ownerId]
      properties:
        id:            { type: integer }
        name:          { type: string }
        targetAmount:  { type: number }
        currentAmount: { type: number }
        targetDate:    { type: ["string", "null"] }
        ownerId:       { type: ["integer", "null"] }
        color:         { type: string }
        icon:          { type: string }
        notes:         { type: ["string", "null"] }
        ownerName:     { type: ["string", "null"] }
        progressPct:   { type: number }
    SavingsGoalInput:
      type: object
      required: [name, targetAmount]
      properties:
        name:          { type: string }
        targetAmount:  { type: number }
        currentAmount: { type: number }
        targetDate:    { type: ["string", "null"] }
        ownerId:       { type: ["integer", "null"] }
        color:         { type: string }
        icon:          { type: string }
        notes:         { type: ["string", "null"] }
    SavingsGoalUpdate:
      type: object
      properties:
        name:          { type: string }
        targetAmount:  { type: number }
        currentAmount: { type: number }
        targetDate:    { type: ["string", "null"] }
        ownerId:       { type: ["integer", "null"] }
        color:         { type: string }
        icon:          { type: string }
        notes:         { type: ["string", "null"] }
    SavingsContribution:
      type: object
      required: [amount]
      properties:
        amount: { type: number }

    # ── Dashboard Analytics ───────────────────────────────────────────────────
...